Fixed sizeEstimate.lo when archiving empty files
In addition another sanity script is provided which
creates directory and sub-directories randomly and also
files. The MD5 sum of all the data is calculated and then
archived. Then the data is locally deleted and retrieved from
the TSM and the MD5 sum is calculated on the retrieved data and
compared with the archived data.