Set default number of consumer threads to 1
TSM server default setting of maximum number of mount points
(short maxnummp) is 1. If this value is not increased, and default
of nthreads in lhsmtool_tsm.c is 2, then this results in
"This node has exceeded its maximum number of mount points" errors.
Thus use same default values on TSM server and lhsmtool_tsm side.