[tsmapi,test] Overwrite objects in hashtable
If newobj has same key (fs/hl/ll) and same date or
later date than the oldobj (which is already in the hashtable),
will be removed, and newobj inserted. This guarantees that
no multiple objects (same key and date) are stored in hashtable.
This fix issue: https://github.com/tstibor/ltsm/issues/20