[tsmapi,copytool] Restore Lustre stripe data
Lustre stripe information are obtained by reading
the xattr value XATTR_LUSTRE_LOV and getting data
lmm_stripe_size and lmm_stripe_count. This patch
enables to restore also the Lustre information
which are stored in TSM objects in the retrieve
operation. If the flag --restore-stripe is not
provided, then the default Lustre information
are used.