[benchmark] Benchmark tool measure archive perf
Measure threaded archive performance. Parameters are:
-z, --size <long> Size of file
-b, --number <int> Number of files to archive
-t, --threads <int> Number of parallel threads perform archiving task

Note, in a first step, --number <N> files are created with random file
names from {"0123456789", "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
"abcdefghijklmnopqrstuvwxyz"}. Each file is filled with random bytes of
--size <S>. Make sure that the maximum number of threads does not exceed
the maximum number of mount points on the TSM server. This benchmark
tool is built when ./configure [...] --enable-tests is set.