[build] Build RPM packages with 'make rpms'
Initial commit for building RPM packages.
If GIT tag has version information X.Y.Z,
then release R=1 is set. If X.Y.Z-R is provided,
then version=X.Y.Z and release=R is substituted.

For Debian packages, when release is not set in
the GIT tag, that is, X.Y.Z is provided only, then
the package is build with version *X.Y.Z.deb in
contrast to RPM name *X.Y.Z-1.rpm. This is conform
with the Debian policy.

Note, the RPM spec file is still work in progress,
the systemd integration still needs to be done.