[ltsmsync] Enable script to check on file size or crc32 checksum
Previous implementation did a crc32 checksum on each file in
query list to verify that file on TSM server and file in Lustre mount
point are equal. Relax this constrain, and trigger retrieving process
when:
* File does not exist,
* File crc32 checksum does not match,
* File size is 0

The default option is to check the file size. Checksum verification
can be set by option, '-c' or '--crc32-verify'.