Set object size in stream/pipe mode to 2GB rather than maximum
In stream/pipe mode the object size is not known a priori, however,
on average archived objects are approx. of size 2GB. Thus set size
to 2GB. Before size was set to maximum (2^64 GB), however, when archiving
process failed e.g. due to network problems, the DSM object size is not
updated afterwards (size is known) and set to 2^64 GB.