Disable polling in STATE_TSM_ARCHIVE_RUN
Previously we stayed in STATE_TSM_ARCHIVE_RUN and polled every 50ms to check,
whether the archive operation finishes successfully. This approach,
however, is not efficient and in addition congests the queue, as the
the fsd_action_item is enqueued/dequeued over and over until the file
is finally archived. To overcome this problem, change immediately
the state from STATE_TSM_ARCHIVE_RUN to STATE_TSM_ARCHIVE_DONE and
assume: If archive_action(fsd_action_item)
returns success, then the file is also successfully archived.