Enable/disable ltsmfsd Lustre HSM state polling via C directive
For making sure that a file is finally archived via the copytool
on the TSM server, one needs to check the extended attribute
flags HS_EXISTS and HS_ARCHIVED. Unfortunately, there seems no
efficient approach to do that, except polling. Defining C directive
LTSMFSD_POLL_ARCHIVE_FINISHED enables the polling mechanism. When
polling is disabled, it is assumed that a successful call
of llapi_hsm_request(...) archives the data on TSM server.