Add FSD_STORAGE_NULL as a storage target.
For testing FSDAPI it is usefull to have a dummy
storage target. The data received via socket with
FSD_STORAGE_NULL as a storage target is received however,
as the name suggests not stored. Such a dummy storage
target is useful for benchmarking plain network connections
and discarding testing data coming from the detectors.