Implemention of FSQ 'REPLY' message
If an error occurred in FSQ server, then
the server socket is closed. As a consequence the client gets
no information on the error reason because no 'REPLY' message
it sent back containing the error code and error str. This
commit introduces in calls fsq_connect and fsq_open the
error code and str message.