Finalize uuid TSM copytool functionality
Archiving a file with TSM copytool on Lustre and subsequently releasing
and well as moving the file to a different directory resulted
in missmatch between the `readlink -f file` and the TSM created object
values /fs/hl/ll. To solve this problem, starting with this commit
each TSM archived Lustre file will have an extended attribute of the
form e.g.: user.lustre.uuid=0xc72e6182ab0b4e00b6798282dea857fe

When the file is moved (or renamed) and the extended attribute
user.lustre.uuid is found, then the TSM query to locate
the file on the TSM server is modified to: /fs/*/* and
description=0xc72e6182ab0b4e00b6798282dea857fe

When no extended attribute user.lustre.uuid is found
the file qualified path of the file (`readlink -f file`)
in the form /fs/hl/ll is used.