Allow archiving when Lustre is not mounted with user_xattr
If Lustre file system is not mounted with option user_xattr and
file is archived and subsequently moved to into another directory
on the Lustre mount point, it cannot be retrieved. For the sake
of compatibility we output an error when setxattr fails, however
continue to archive the file without UUID xattr.